refactor(island): element renamed IslandTemplate and update snapshots
