+++ libs/island/css.ts
@@ -1,22 +1,43 @@
-// -- JSS --
-import { create as createJSS, type Styles } from 'https://esm.sh/jss@10.9.2'
-import { default as nestedJSS } from 'https://esm.sh/jss-plugin-nested@10.9.2'
-import { default as camelJSS } from 'https://esm.sh/jss-plugin-camel-case@10.9.2'
-import { default as globalJSS } from 'https://esm.sh/jss-plugin-global@10.9.2'
+// import { parsel } from './deps.ts'
+// import { taggedWithPrimitives } from './utils.ts'
+// import { hashString } from '../utils/hash.ts'
+// type Primitive =
+//   | number
+//   | string
+//   | boolean
+//   | undefined
+//   | null
+//   | void
 
-export const css = (style: Styles) => {
-  const jss = createJSS()
-  jss.use(
-    globalJSS(),
-    nestedJSS(),
-    camelJSS(),
-  )
-  const sheet = jss.createStyleSheet(style)
-  const stylesheet = sheet.toString()
-  return {
-    styles: sheet.classes,
-    stylesheet,
-  }
-}
-
-export { type Styles }
+// export const css = (
+//   strings: TemplateStringsArray,
+//   ...expressions: Array<Primitive | Primitive[]>
+// ) => {
+//   const result = taggedWithPrimitives(strings, ...expressions)
+//   const prefix = btoa(`${hashString(result)}`).replace(/[+/=]/g, '')
+//   const parser = parsel
+//   const { type, ...rest } = parsel.TOKENS
+//   const TOKENS = {
+//     ...rest,
+//     nesting: /&/g,
+//     nest: /@nest\b/g,
+//     type,
+//   }
+//   parser.TOKENS = TOKENS
+//   const tokens = parser.tokenize(result)
+//   // const { transformed } = tokens.reduce(
+//   //   () => {},
+//   //   {
+//   //     lastRawSelector: undefined,
+//   //     lastTransformedSelector: undefined,
+//   //     lastOpenBracket: undefined,
+//   //     lastCloseBracket: undefined,
+//   //     raw: [],
+//   //     transformed: [],
+//   //   },
+//   // )
+//   return {
+//     styles: {},
+//     stylesheet: '',
+//   }
+// }
