+++ libs/island/deps.ts
@@ -1,0 +1,8 @@
+// // -- lodash --
+// export {
+//   camelCase,
+//   kebabCase,
+// } from 'https://raw.githubusercontent.com/lodash/lodash/4.17.21-es/lodash.js'
+
+// -- parsel --
+export * as parsel from 'https://esm.sh/parsel-js@1.0.3'
