+++ libs/island/html.ts
@@ -1,39 +1,12 @@
-import { escape, trueTypeOf } from '../utils/mod.ts'
-
-type Primitive =
-  | number
-  | string
-  | boolean
-  | undefined
-  | null
-  | void
-
-const reduceWhitespace = (str: string) => str.replace(/(\s\s+|\n)/g, ' ')
-
-const isTruthy = (val: Primitive) =>
-  trueTypeOf(val) === 'string' ||
-  trueTypeOf(val) === 'number'
+import { taggedWithPrimitives } from './utils.ts'
+import { Primitive } from './types.ts'
 
 /** @returns a minimized html template string */
 export const html = (
   strings: TemplateStringsArray,
   ...expressions: Array<Primitive | Primitive[]>
 ) => {
-  const { raw } = strings
-  let result = expressions.reduce<string>((acc, subst, i) => {
-    acc += reduceWhitespace(raw[i])
-    let filteredSubst = Array.isArray(subst)
-      ? subst.filter(isTruthy).join('')
-      : isTruthy(subst)
-      ? subst
-      : ''
-    if (acc.endsWith('$')) {
-      filteredSubst = escape(filteredSubst as string)
-      acc = acc.slice(0, -1)
-    }
-    return acc + filteredSubst
-  }, '')
-  result += reduceWhitespace(raw[raw.length - 1])
+  const result = taggedWithPrimitives(strings, ...expressions)
   const tpl = result
     .trim()
     .replace(/[\s\n]+>/g, '>')
