+++ libs/island/mod.ts
@@ -1,5 +1,5 @@
 export * from './class-names.ts'
-// export * from './css.ts'
+export * from './css.ts'
 export * from './css-var.ts'
 export * from './constants.ts'
 export * from './island.ts'
